% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.panel.r
\name{build.psid}
\alias{build.psid}
\title{Build full example PSID}
\usage{
build.psid()
}
\value{
a list with panel data and data dictionary
}
\description{
Builds a panel from the full PSID dataset (i.e. all years)
}

