\name{get.psid}
\alias{get.psid}
\title{get.psid connects to PSID database and downloads into Rda}
\usage{
  get.psid(file, name, params, curl)
}
\description{
  see \url{http://www.asdfree.com/} for other usage and
  \url{http://stackoverflow.com/questions/15853204/how-to-login-and-then-download-a-file-from-aspx-web-pages-with-r}
}
\author{
  Anthony Damico <ajdamico@gmail.com>
}

