% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeids.r
\name{make.char}
\alias{make.char}
\title{Convert factor to character}
\usage{
make.char(x)
}
\arguments{
\item{x}{a \code{factor}}
}
\value{
a character
}
\description{
helper function to convert factor to character in a data.table
}
