% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.panel.r
\name{medium.test.ind.NA.wealth}
\alias{medium.test.ind.NA.wealth}
\title{three year test, ind file and one NA variable and wealth}
\usage{
medium.test.ind.NA.wealth(dd = NULL)
}
\description{
three year test, ind file and one NA variable and wealth
}
