% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.panel.r
\name{medium.test.ind}
\alias{medium.test.ind}
\title{three year test, ind file}
\usage{
medium.test.ind(dd = NULL)
}
\arguments{
\item{dd}{Data Dictionary location. If NULL, 
use temp dir and force download}
}
\description{
three year test, ind file
}
