% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.panel.r
\name{medium.test.noind}
\alias{medium.test.noind}
\title{three year test, no ind file}
\usage{
medium.test.noind(dd = NULL)
}
\arguments{
\item{dd}{Data Dictionary location}
}
\description{
three year test, no ind file
}
