% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.panel.r
\name{small.test.ind}
\alias{small.test.ind}
\title{one year test, ind file}
\usage{
small.test.ind(dd = NULL)
}
\arguments{
\item{dd}{Data Dictionary location. If NULL, 
use temp dir and force download}
}
\value{
No return value, called for side effects
}
\description{
one year test, ind file
}
