% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeids.r
\name{testPSID}
\alias{testPSID}
\title{Create a test PSID dataset}
\usage{
testPSID(N = 100, N.attr = 0)
}
\arguments{
\item{N}{number of people in each wave}

\item{N.attr}{number of people lost to attrition}
}
\value{
list with (fake) individual index file IND2009ER and
family files for 1985 and 1986
}
\description{
makes artifical PSID data with variables \code{age} and \code{income}
for two consecutive years 1985 and 1986.
}

