% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.panel.r
\name{build.panel}
\alias{build.panel}
\title{build.panel: Build PSID panel data set}
\usage{
build.panel(datadir = NULL, fam.vars, ind.vars = NULL, wealth.vars = NULL,
  SAScii = FALSE, heads.only = FALSE, current.heads.only = FALSE,
  sample = NULL, design = "balanced", verbose = FALSE)
}
\arguments{
\item{datadir}{either \code{NULL}, in which case saves to tmpdir or path to directory containing family files ("FAMyyyy.xyz") and individual file ("IND2009ER.xyz") in admissible formats .xyz. Admissible are .dta, .csv, .RData, .rda. Please follow naming convention. Only .dta version <= 12 supported. Recommended usage is to specify \code{datadir}.}

\item{fam.vars}{data.frame of variable to retrieve from family files. Can contain see example for required format.}

\item{ind.vars}{data.frame of variables to get from individual file. In almost all cases this will be the type of survey weights you want to use. don't include id variables ER30001 and ER30002.}

\item{wealth.vars}{data.frame of variables to get from the wealth supplement files.}

\item{SAScii}{logical TRUE if you want to directly download data into Rda format (no dependency on STATA/SAS/SPSS). may take a long time, but downloads only once if you specify \code{datadir}.}

\item{heads.only}{logical TRUE if user wants household heads only. Household heads in sample year.}

\item{current.heads.only}{logical TRUE if user wants current household heads only. Distinguishes mover outs heads.}

\item{sample}{string indicating which sample to select: "SRC" (survey research center), "SEO" (survey for economic opportunity), "immigrant" (immigrant sample), "latino" (Latino family sample). Defaults to NULL, so no subsetting takes place.}

\item{design}{either character \emph{balanced} or \emph{all} or integer. \emph{balanced} means only individuals who appear in each wave are considered. \emph{All} means all are taken. An integer value stands for minimum consecutive years of participation, i.e. design=3 means present in at least 3 consecutive waves.}

\item{verbose}{logical TRUE if you want verbose output.}
}
\value{
\item{data}{resulting \code{data.table}. the variable \code{pid} is the unique person identifier, constructed from ID1968 and pernum.}
\item{dict}{data dictionary if stata data was supplied, NULL else}
}
\description{
Builds a panel data set with id variables \code{pid} (unique person identifier) and \code{year} from individual PSID family files and supplemental wealth files.
}
\details{
There are several supported approches. Approach one downloads stata data, uses stata to build each wave, then puts it together with `psidR`. The second (recommended) approach downloads all data directly from the psid servers (no Stata needed. For this approach you need to supply the precise names of psid variables - those variable names vary by year. E.g. \emph{total family income} will have different names in different waves. The function \code{\link{getNamesPSID}} greatly helps collecting names for all waves.
Merge: the variables \code{interview number} in each family file map to 
the \code{interview number} variable of a given year in the individual file. Run \code{example(build.panel)} for a demonstration.
For approach one, accepted input data are stata format .dta, .csv files or R data formats .rda and RData. This usage is similar to stata module \code{psiduse}. 
Approach two follows the strategy introduced at \url{http://asdfree.com}. In fact, both approaches use the same function \code{save.psid} to download the data, \code{psidR} automates the merge and subsetting proceedure for you.
}
\examples{
\dontrun{
# ################################################
# Real-world example: not run because takes long.
# Build panel with income, wage, age and education
# optionally: add wealth supplements!
# ################################################

# The package is installed with a list of variables
# Alternatively, search for names with \\code{\\link{getNamesPSID}}
# This is the body of function build.psid()
# (so why not call build.psid() and see what happens!)
r = system.file(package="psidR")
if (small){
  f = fread(file.path(r,"psid-lists","famvars-small.txt"))
  i = fread(file.path(r,"psid-lists","indvars-small.txt"))
  if (wealth){
    w = fread(file.path(r,"psid-lists","wealthvars-small.txt"))
  }
} else {
  f = fread(file.path(r,"psid-lists","famvars.txt"))
  i = fread(file.path(r,"psid-lists","indvars.txt"))
  if (wealth){
    w = fread(file.path(r,"psid-lists","wealthvars.txt"))
  }
}
setkey(i,"name")
setkey(f,"name")
if (wealth) setkey(w,"name")
i = dcast(i[,list(year,name,variable)],year~name)
f = dcast(f[,list(year,name,variable)],year~name)
if (wealth) {
  w = dcast(w[,list(year,name,variable)],year~name)
  d = build.panel(datadir="~/datasets/psid/",fam.vars=f,
                 ind.vars=i,wealth.vars=w,SAScii = TRUE, 
                 heads.only =TRUE,sample="SRC",verbose=TRUE,
                 design="all")
  save(d,file="~/psid_wealth.RData")
} else {
  d = build.panel(datadir="~/datasets/psid/",fam.vars=f,
                 ind.vars=i,SAScii = TRUE, 
                 heads.only =TRUE,sample="SRC",verbose=TRUE,
                 design="all")
  save(d,file="~/psid_no_wealth.RData")
}
}

# ######################################
# reproducible example on artifical data. 
# run this with example(build.panel).
# ######################################

## make reproducible family data sets for 2 years
## variables are: family income (Money) and age

## Data acquisition step: you download data or
## run build.panel with sascii=TRUE

# testPSID creates artifical PSID data
td <- testPSID(N=12,N.attr=0)
fam1985 <- data.table::copy(td$famvars1985)
fam1986 <- data.table::copy(td$famvars1986)
IND2009ER <- data.table::copy(td$IND2009ER)

# create a temporary datadir
my.dir <- tempdir()
#save those in the datadir
# notice different R formats admissible
save(fam1985,file=paste0(my.dir,"/FAM1985ER.rda"))
save(fam1986,file=paste0(my.dir,"/FAM1986ER.RData"))
save(IND2009ER,file=paste0(my.dir,"/IND2009ER.RData"))

## end Data acquisition step.

# now define which famvars
famvars <- data.frame(year=c(1985,1986),
                      money=c("Money85","Money86"),
                      age=c("age85","age86"))

# create ind.vars
indvars <- data.frame(year=c(1985,1986),ind.weight=c("ER30497","ER30534"))

# call the builder
# data will contain column "relation.head" holding the relationship code.

d <- build.panel(datadir=my.dir,fam.vars=famvars,
                 ind.vars=indvars,
                 heads.only=FALSE,verbose=TRUE)	

# see what happens if we drop non-heads
# only the ones who are heads in BOTH years 
# are present (since design='balanced' by default)
d <- build.panel(datadir=my.dir,fam.vars=famvars,
                 ind.vars=indvars,
                 heads.only=TRUE,verbose=FALSE)	
print(d$data[order(pid)],nrow=Inf)

# change sample design to "all": 
# we'll keep individuals if they are head in one year,
# and drop in the other
d <- build.panel(datadir=my.dir,fam.vars=famvars,
                 ind.vars=indvars,heads.only=TRUE,
                 verbose=FALSE,design="all")	
print(d$data[order(pid)],nrow=Inf)

file.remove(paste0(my.dir,"/FAM1985ER.rda"),
            paste0(my.dir,"/FAM1986ER.RData"),
            paste0(my.dir,"/IND2009ER.RData"))

# END psidR example

# #####################################################################
# Please go to https://github.com/floswald/psidR for more example usage
# #####################################################################
}
