% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.panel.r
\name{build.psid}
\alias{build.psid}
\title{Build example PSID}
\usage{
build.psid(datadr = "~/datasets/psid/", small = TRUE)
}
\arguments{
\item{datadr}{string of the data directory}

\item{small}{logical TRUE if only use years 2013 and 2015.}
}
\value{
a data.table with panel data
}
\description{
Builds a panel from the full PSID dataset
}
