% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.panel.r
\name{medium.test.ind.NA}
\alias{medium.test.ind.NA}
\title{three year test, ind file and one NA variable}
\usage{
medium.test.ind.NA(dd = NULL)
}
\arguments{
\item{dd}{Data Dictionary location. If NULL, 
use temp dir and force download}
}
\value{
No return value, called for side effects
}
\description{
three year test, ind file and one NA variable
}
